+++ src/types/Database.ts
@@ -15,11 +15,20 @@ export interface IDatabaseConfiguration {
     collectionName?: string
 }
 
+/**
+ * Represents the object of the way data stored in the internal
+ * `[__KEY]-[__VALUE]` storage format that was made to achieve better data accessibility across the module.
+ *
+ * @template T The type of `__VALUE` property in each raw data object.
+ */
 export interface IDatabaseInternalStructure<T = any> {
     __KEY: string
     __VALUE: T
 }
 
+/**
+ * Represents the database operations latency object.
+ */
 export type IDatabaseRequestsLatencyData = Record<
     'readLatency' | 'writeLatency' | 'deleteLatency',
     number