+++ tests/errors.test.ts
@@ -1,17 +1,21 @@
 /* eslint-disable no-return-await */
 
-import { promisify } from 'util'
-
 import { describe, expect, test, afterAll, beforeAll } from '@jest/globals'
 import { QuickMongo, QuickMongoClient } from '../src'
 
-const sleep = promisify(setTimeout)
-
 let quickMongoClient = new QuickMongoClient('mongodb://127.0.0.1:27018')
+let database: QuickMongo<string, any>
 
 beforeAll(async () => {
     const client = await quickMongoClient.connect()
     quickMongoClient = client
+
+    database = new QuickMongo<string, any>(quickMongoClient, {
+        name: 'test_database',
+        collectionName: 'test_database_collection'
+    })
+
+    await database.deleteAll()
 })
 
 const resolvePromise = (promise: Promise<any>): Promise<boolean> => {
@@ -35,11 +39,6 @@ const resolveFunction = (fn: (...args: any[]) => any): boolean => {
 // getting data
 
 describe('errors throwing: fetch()', () => {
-    const database = new QuickMongo<string, any>(quickMongoClient, {
-        name: 'test_database',
-        collectionName: 'test_database_collection'
-    })
-
     test.concurrent('\'key\' parameter is missing', async () => {
         // @ts-expect-error
         const errorTest = (): any => database.fetch()
@@ -53,9 +52,6 @@ describe('errors throwing: fetch()', () => {
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successTest = (): any => database.fetch('test')
         return expect(resolveFunction(successTest)).toBeTruthy()
     })
@@ -80,9 +76,6 @@ describe('errors throwing: has()', () => {
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successCase = (): any => database.has('test')
         return expect(resolveFunction(successCase)).toBeTruthy()
     })
@@ -101,17 +94,11 @@ describe('errors throwing: keys()', () => {
     })
 
     test.concurrent('success case: \'key\' parameter is missing (return database root keys)', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successCase = (): any => database.keys()
         return expect(resolveFunction(successCase)).toBeTruthy()
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successCase = (): any => database.keys('test')
         return expect(resolveFunction(successCase)).toBeTruthy()
     })
@@ -145,9 +132,6 @@ describe('errors throwing: set()', () => {
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successCase = async (): Promise<any> => await database.set('test', 123)
         return expect(resolvePromise(successCase())).resolves.toBeTruthy()
     })
@@ -172,9 +156,6 @@ describe('errors throwing: delete()', () => {
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successCase = async (): Promise<any> => await database.delete('test')
         return expect(resolvePromise(successCase())).resolves.toBeTruthy()
     })
@@ -224,9 +205,6 @@ describe('errors throwing: add()', () => {
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         await database.set('arr', [])
 
         const successTest = async (): Promise<any> => await database.add('test', 123)
@@ -278,9 +256,6 @@ describe('errors throwing: subtract()', () => {
     })
 
     test.concurrent('success case', async () => {
-        await database.loadCache()
-        await sleep(1000)
-
         const successTest = async (): Promise<any> => await database.subtract('test', 123)
         return expect(resolvePromise(successTest())).resolves.toBeTruthy()
     })
@@ -288,9 +263,11 @@ describe('errors throwing: subtract()', () => {
 
 
 // array methods
+// soon...
+
 
 // post-testing cleanup
+
 afterAll(async () => {
-    await Promise.all(quickMongoClient.databases.map(database => database.deleteAll()))
     await quickMongoClient.disconnect()
 })